Rename @export to @EntryPoint and add reachability assertions.
