+++ compiler/src/dotty/tools/dotc/core/Definitions.scala
@@ -515,8 +515,8 @@ class Definitions {
   def ContravariantBetweenAnnot(implicit ctx: Context) = ContravariantBetweenAnnotType.symbol.asClass
   lazy val DeprecatedAnnotType = ctx.requiredClassRef("scala.deprecated")
   def DeprecatedAnnot(implicit ctx: Context) = DeprecatedAnnotType.symbol.asClass
-  lazy val ExportAnnotType = ctx.requiredClassRef("scala.export")
-  def ExportAnnot(implicit ctx: Context) = ExportAnnotType.symbol.asClass
+  lazy val EntryPointAnnotType = ctx.requiredClassRef("scala.EntryPoint")
+  def EntryPointAnnot(implicit ctx: Context) = EntryPointAnnotType.symbol.asClass
   lazy val ImplicitNotFoundAnnotType = ctx.requiredClassRef("scala.annotation.implicitNotFound")
   def ImplicitNotFoundAnnot(implicit ctx: Context) = ImplicitNotFoundAnnotType.symbol.asClass
   lazy val InlineAnnotType = ctx.requiredClassRef("scala.inline")
