+++ compiler/test/dotc/tests.scala
@@ -208,7 +208,7 @@ class tests extends CompilerTest {
   // Test callgraph DCE on code that use DCEed stdlib
   @Test def link_stdlib_dce_all =
     runFiles(linkDir + "stdlib-dce/", List("-language:Scala2", "-link-dce"), stdlibFiles = stdlibFiles)
-  @Test def link_stdlib_dce_vis_all =
+  @org.junit.Ignore("Takes too long in 'sbt test'") @Test def link_stdlib_dce_vis_all =
     runFiles(linkDir + "stdlib-dce/", List("-language:Scala2", "-link-dce", "-link-vis"), stdlibFiles = stdlibFiles)
 
   private val stdlibFiles: List[String] = Source.fromFile("./test/dotc/scala-collections.whitelist", "UTF8").getLines()
