+++ library/src/scala/EntryPoint.scala
@@ -1,0 +1,3 @@
+package scala
+
+final class EntryPoint extends scala.annotation.StaticAnnotation
