--- library/src/scala/annotation/internal/DoNotDCE.scala
@@ -1,3 +1,0 @@
-package scala.annotation.internal
-
-final class DoNotDCE extends scala.annotation.StaticAnnotation
