+++ library/src/scala/annotation/internal/link/AssertNotReachable.scala
@@ -1,0 +1,8 @@
+package scala.annotation.internal.link
+
+/** Assert at compile time that the symbol of a method is not reachable.
+  *
+  * This does not work on methods that get created later like getters,
+  * setters, bridges, ...
+  */
+final class AssertNotReachable extends scala.annotation.StaticAnnotation
