+++ library/src/scala/annotation/internal/link/DoNotDeadCodeEliminate.scala
@@ -1,0 +1,4 @@
+package scala.annotation.internal.link
+
+/** Ignores dead code elimination of this method */
+final class DoNotDeadCodeEliminate extends scala.annotation.StaticAnnotation
