--- library/src/scala/export.scala
@@ -1,3 +1,0 @@
-package scala
-
-final class export extends scala.annotation.StaticAnnotation
