+++ tests/link/dce-failing/link-code-from-java-6.check
