--- tests/link/dce-failing/link-exports-3/DCETest.scala
@@ -1,15 +1,0 @@
-
-object DCETest {
-  @scala.annotation.internal.DoNotDCE def dceTest: Unit = {
-    Foo.bar()
-    Foo.foo()
-  }
-}
-
-object Foo {
-
-  bar()
-
-  @scala.export def foo(): Unit = System.out.println(42)
-  def bar(): Unit = System.out.println(43)
-}
