+++ tests/link/dce-failing/link-main-module-load.scala
@@ -1,9 +1,10 @@
+import scala.annotation.internal
 
 object Test {
 
   def main(args: Array[String]): Unit = ()
 
-  def foo() = System.out.println(42)
+  @internal.link.AssertReachable def foo() = System.out.println(42)
 
   foo()
