+++ tests/link/dce/link-code-from-java-1.scala
@@ -1,7 +1,37 @@
-import scala.annotation.internal.DoNotDCE
+import java.io.File
+import java.io.FileFilter
+import java.lang.reflect.InvocationTargetException
+import java.lang.reflect.Method
+import java.lang.reflect.Type
 
-object DCETest {
-  @DoNotDCE def dceTest: Unit = {
+import scala.annotation.internal
+
+object Test {
+  def main(args: Array[String]): Unit = {
+    val classLoader = Test.getClass.getClassLoader()
+
+    try {
+      val mainClass = classLoader.loadClass("Test")
+      val mainMethod = mainClass.getMethod("dceTest")
+      mainMethod.invoke(null);
+    } catch {
+      case e: java.lang.Exception => e.getCause.printStackTrace()
+    }
+  }
+
+  @internal.link.AssertNotReachable
+  @internal.link.DoNotDeadCodeEliminate
+  def shouldDCE(expr: => Any): Unit = try {
+    expr
+    throw new Exception("Expected DCE")
+  } catch {
+    case dce: dotty.runtime.DeadCodeEliminated =>
+    // TODO: check stack trace to see if the DCE was in the fist call of expr
+  }
+
+  @internal.link.AssertNotReachable
+  @internal.link.DoNotDeadCodeEliminate
+  def dceTest: Unit = {
     System.out.println("dceTest")
 
     val a = new A
@@ -23,7 +53,7 @@ object DCETest {
     c.toString
   }
 
-  @scala.export def entryPoint(): Unit = {
+  @EntryPoint def entryPoint(): Unit = {
     System.out.println(new A)
     System.out.println(new B {})
     System.out.println(new C {})
