+++ tests/link/dce/link-code-from-java-2.scala
@@ -1,9 +1,33 @@
 import java.util.Observable
 
-import scala.annotation.internal.DoNotDCE
+import scala.annotation.internal
 
-object DCETest {
-  @DoNotDCE def dceTest: Unit = {
+object Test {
+  def main(args: Array[String]): Unit = {
+    val classLoader = Test.getClass.getClassLoader()
+
+    try {
+      val mainClass = classLoader.loadClass("Test")
+      val mainMethod = mainClass.getMethod("dceTest")
+      mainMethod.invoke(null);
+    } catch {
+      case e: java.lang.Exception => e.getCause.printStackTrace()
+    }
+  }
+
+  @internal.link.AssertNotReachable
+  @internal.link.DoNotDeadCodeEliminate
+  def shouldDCE(expr: => Any): Unit = try {
+    expr
+    throw new Exception("Expected DCE")
+  } catch {
+    case dce: dotty.runtime.DeadCodeEliminated =>
+    // TODO: check stack trace to see if the DCE was in the fist call of expr
+  }
+
+  @internal.link.AssertNotReachable
+  @internal.link.DoNotDeadCodeEliminate
+  def dceTest: Unit = {
     System.out.println("dceTest")
 
     val a = new A
@@ -28,7 +52,7 @@ object DCETest {
     c.update(new Observable, null)
   }
 
-  @scala.export def entryPoint(): Unit = {
+  @EntryPoint def entryPoint(): Unit = {
     System.out.print(new A)
     System.out.print(new B {})
     System.out.print(new C {})
