+++ tests/link/dce/link-codeCalledFrom-2.scala
@@ -1,3 +1,4 @@
+import scala.annotation.internal
 
 object Test {
   def main(args: Array[String]): Unit = {
@@ -11,5 +12,6 @@ object Test {
 }
 
 class Foo {
+  @internal.link.AssertReachable
   override def toString: String = "Foo.toString"
 }
