+++ tests/link/dce/link-constructorParams.scala
@@ -1,4 +1,4 @@
-
+import scala.annotation.internal
 
 object Test {
   def main(args: Array[String]): Unit = {
@@ -7,4 +7,4 @@ object Test {
 }
 
 class Foo(x: Int) extends Bar(x)
-class Bar(val x: Int)
+class Bar(@internal.link.AssertReachable val x: Int)
