+++ tests/link/dce/link-entry-point-1.check
