+++ tests/link/dce/link-entry-point-2.check
