+++ tests/link/dce/link-entry-point-4.check
