+++ tests/link/dce/link-innerClass-0.scala
@@ -1,7 +1,9 @@
+import scala.annotation.internal
 
 object Test {
   def main(args: Array[String]): Unit = {
     class Foo {
+      @internal.link.AssertReachable
       def bar: Int = 42
     }
