+++ tests/link/dce/link-innerClass-1.scala
@@ -1,3 +1,4 @@
+import scala.annotation.internal
 
 object Test {
   def main(args: Array[String]): Unit = {
@@ -8,6 +9,7 @@ object Test {
 class Bar {
   def test() = {
     class Foo {
+      @internal.link.AssertReachable
       def bar: Int = 42
     }
