+++ tests/link/dce/link-val-init-1.scala
@@ -1,5 +1,7 @@
+import scala.annotation.internal
 
 class Foo {
+  @internal.link.AssertReachable
   def foo = System.out.println(42)
   val bar = foo
 }
