+++ tests/link/dce/stdlib-link-mutable-set-0/stdlib-link-mutable-set-0.scala
