+++ tests/link/dce/stdlib-link-predef-println-1/stdlib-link-predef-println-1.scala
