+++ tests/link/dce/stdlib-link-predef-println-2/stdlib-link-predef-println-2.scala
